pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]

[project]
name = "affiliate-marketing-automation"
version = "0.1.0"
description = "AI-powered affiliate marketing automation for Vietnamese e-commerce platforms"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "arq>=0.26.0",
    "redis>=5.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10.0",
    "pillow>=10.0.0",
    "tenacity>=8.0.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
